requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pandas==2.2.3
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
portalocker==2.8.0
python-multipart==0.0.6
pydantic==1.10.12
